backend/requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
pydantic==1.10.11
APScheduler==3.10.1
email-validator==1.3.1
requests==2.31.0
gunicorn==21.2.0
python-dotenv==1.0.0
